Moved Helseatlas front page code into component
More sharing of code between english and norwegian page.
