+++ apps/skde/src/components/Atlas/FrontPage/FrontPage.module.css
@@ -1,0 +1,9 @@
+.atlasLinksWrapper {
+  display: flex;
+  justify-content: space-between;
+  flex-wrap: wrap;
+  margin: 40px auto;
+  padding-left: 16px;
+  padding-right: 16px;
+  max-width: min(1216px, 95%);
+}
