+++ apps/skde/src/components/MainBanner/MainBanner.tsx
@@ -4,21 +4,21 @@ import { imgLoader } from "../../helpers/functions";
 import classNames from "./MainBanner.module.css";
 
 type MainBannerProps = {
-  lang?: "nb" | "nn" | "en";
-  title?: string;
-  ingress?: string;
-  imageSource?: string;
+  lang: "no" | "en";
 };
 
-export const MainBanner: React.FC<MainBannerProps> = ({
-  lang = "no",
-  title = "Likeverdige helsetjenester – uansett hvor du bor?",
-  ingress = "Helseatlas sammenlikner befolkningens bruk av helsetjenester i forskjellige geografiske områder, uavhengig av hvilket sted pasientene behandles.",
-  imageSource = "/helseatlas/img/frontpage.jpg",
-}) => {
-  const engTitle = "Equitable health services – regardless of where you live?";
-  const engIngress =
-    "In Norway, it is a goal for the entire population to have an equal supply of health services across geography and social groups. The health atlas is a tool for comparing the population's use of health services in different geographical areas, regardless of where the patients are treated.";
+export const MainBanner: React.FC<MainBannerProps> = ({ lang }) => {
+  const text = {
+    title: {
+      en: "Equitable health services – regardless of where you live?",
+      no: "Likeverdige helsetjenester – uansett hvor du bor?",
+    },
+    ingress: {
+      en: "In Norway, it is a goal for the entire population to have an equal supply of health services across geography and social groups. The health atlas is a tool for comparing the population's use of health services in different geographical areas, regardless of where the patients are treated.",
+      no: "Helseatlas sammenlikner befolkningens bruk av helsetjenester i forskjellige geografiske områder, uavhengig av hvilket sted pasientene behandles.",
+    },
+  };
+  const imageSource = "/helseatlas/img/frontpage.jpg";
   return (
     <div className={classNames.main}>
       <div className={classNames.headerMain}>
@@ -30,16 +30,12 @@ export const MainBanner: React.FC<MainBannerProps> = ({
             width={1184}
             height={435}
           />
-          <h1 className={classNames.pageTitle}>
-            {lang === "en" ? engTitle : title}
-          </h1>
+          <h1 className={classNames.pageTitle}>{text.title[lang]}</h1>
         </div>
       </div>
       <div className={classNames.pageTitle}></div>
       <div className={classNames.pageIngressContent}>
-        <div className={classNames.pageIngress}>
-          {lang === "en" ? engIngress : ingress}
-        </div>
+        <div className={classNames.pageIngress}>{text.ingress[lang]}</div>
       </div>
     </div>
   );
